#70 Refactored SwagDtoProperty and SwagForm
- Split SwagDtoProperty into new annotations: SwagDtoQuery & SwagDtoForm
- Extend SwagDtoForm, SwagForm, and SwagEntityAttribute from
AbstractSchemaProperty
- Refactored DtoParser to support changes
- Extend SwagDto from AbstractParameter
- Updated README
- Updated unit tests
